% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_keras_model.R
\name{build_keras_model}
\alias{build_keras_model}
\title{Title}
\usage{
build_keras_model(p, af_string_list, h_neurons_vector, my_max_norm)
}
\arguments{
\item{p}{a}

\item{af_string_list}{a}

\item{h_neurons_vector}{a}

\item{my_max_norm}{a}
}
\value{
a
}
\description{
Title
}
