% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_l1_norm.R
\name{constraint_l1_norm}
\alias{constraint_l1_norm}
\title{Define a custom keras constraint}
\usage{
constraint_l1_norm(w)
}
\arguments{
\item{w}{The weights to which the constraint will be applied}
}
\value{
Constrained weights
}
\description{
Define a custom keras constraint
}
