% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_train_test.R
\name{divide_train_test}
\alias{divide_train_test}
\title{Divide data in train and test}
\usage{
divide_train_test(data, train_proportion)
}
\arguments{
\item{data}{input data}

\item{train_proportion}{\% of train}
}
\value{
list with train and test
}
\description{
Divide data in train and test
}
