% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_normal_data.R
\name{generate_normal_data}
\alias{generate_normal_data}
\title{Generate normal data (X) and a polynomial response (Y) for testing purposes}
\usage{
generate_normal_data(
  n_sample,
  p,
  q_original,
  mean_range,
  beta_range,
  error_var
)
}
\arguments{
\item{n_sample}{number of samples}

\item{p}{dimensions}

\item{q_original}{degree of the original polynomial}

\item{mean_range}{range for the normal mean}

\item{beta_range}{range for the ebta coefficients}

\item{error_var}{variance of the added normal error}
}
\value{
list with data and orginal coefficients
}
\description{
Generate normal data (X) and a polynomial response (Y) for testing purposes
}
