% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_uniform_data.R
\name{generate_uniform_data}
\alias{generate_uniform_data}
\title{Generate uniform data (X) and a polynomial response (Y) for testing purposes}
\usage{
generate_uniform_data(n_sample, p, q_original, unif_range, error_var)
}
\arguments{
\item{n_sample}{number of samples}

\item{p}{dimensions}

\item{q_original}{degree of the original polynomial}

\item{unif_range}{range for the uniform}

\item{error_var}{variance of the added normal error}
}
\value{
list with data and orginal coefficients
}
\description{
Generate uniform data (X) and a polynomial response (Y) for testing purposes
}
