% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimizer_name_as_string.R
\name{get_optimizer_name_as_string}
\alias{get_optimizer_name_as_string}
\title{Get optimizer name as string}
\usage{
get_optimizer_name_as_string(my_optimizer)
}
\arguments{
\item{my_optimizer}{a keras optimizer function}
}
\value{
string
}
\description{
Get optimizer name as string
}
