% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_coeffs_from_polyreg.R
\name{rename_coeffs_from_polyreg}
\alias{rename_coeffs_from_polyreg}
\title{Renames the coefficients from polyreg package to match our notation}
\usage{
rename_coeffs_from_polyreg(coeffs_polyreg)
}
\arguments{
\item{coeffs_polyreg}{vector with the coefficients of a PR obtained with polyreg}
}
\value{
a vector with the same coeffcients, renamed
}
\description{
Renames the coefficients from polyreg package to match our notation
}
