% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Function that performs one of the following three methods to scale the given data:
\itemize{
\item Scale to the (0,1) interval -> key: "0,1"
\item Scale to the (-1,1) interval -> key: "-1,1"
\item Standardizing the data (mean=0 and sd=1) -> key: "standardize"
}}
\usage{
scale_data(data, scale_method = "0,1")
}
\arguments{
\item{data}{data used}

\item{scale_method}{method employed}
}
\value{
data scaled
}
\description{
Function that performs one of the following three methods to scale the given data:
\itemize{
\item Scale to the (0,1) interval -> key: "0,1"
\item Scale to the (-1,1) interval -> key: "-1,1"
\item Standardizing the data (mean=0 and sd=1) -> key: "standardize"
}
}
